Normalize climate sensor initialization errors
Document shared ClimateSensor begin() result codes and switch the
interface to byte-sized return values. Map SimpleDHT errors to the
shared ClimateSensor codes instead of returning library-specific values.

+++ corsa-a-clock.ino
@@ -68,7 +68,7 @@ void setup()
   buttonTab.begin();
   buttonEnter.begin();
   byte rtcInitializationResult = rtc.begin();
-  int internalClimateSensorInitializationError = internalClimateSensor.begin();
+  byte internalClimateSensorInitializationError = internalClimateSensor.begin();
   {
     char errorValue[12];
     if (rtcInitializationResult != 0)
